+++ src/Utils/index.js
@@ -1,12 +1,5 @@
 export { discountCalculator } from "./discountCalculator";
-export { useDebouncing } from "./debouncing";
-export { useQuery } from "./query";
+export { useDebouncing } from "./Debouncing";
+export { useQuery } from "./Query";
 export { useRequest } from "./request";
-export const getProductWithFlags = ({ cartlist, wishlist, product }) => {
-  const productsIdInCart = cartlist.map((item) => item.details._id);
-  const productsIdInWishList = wishlist.map((item) => item.details._id);
-  return {
-    inCartlist: productsIdInCart.includes(product?._id),
-    inWishlist: productsIdInWishList.includes(product?._id),
-  };
-};
+export { getProductWithFlags } from "./getProductWithFlags";
